Add basic event for emitting DNS dynamic update data
This also changes the existing DU btest from checking to ensure a weird was
emitted to checking the output from the event.

+++ src/analyzer/protocol/dns/DNS.h
@@ -337,7 +337,7 @@ public:
 
     StringValPtr query_name;
     RR_Type atype = TYPE_ALL;
-    int aclass = 0; ///< normally = 1, inet
+    uint16_t aclass = 0; ///< normally = 1, inet
     uint32_t ttl = 0;
 
     DNS_AnswerType answer_type = DNS_QUESTION;
